Add netstandard2.1 target with conditional intrinsic paths
Brings back reach for .NET Core 3.x, Mono, Xamarin and Unity 2021+, which
Tedd.SpanUtils needs before it can reference this package. .NET Framework
does not support netstandard2.1 and still needs the 1.x line.

Neither System.Numerics.BitOperations nor System.Runtime.Intrinsics exists
on netstandard2.1 (both arrived in .NET Core 3.0), so every use of them is
now behind "#if NET6_0_OR_GREATER". Modern targets keep exactly the code
they had; netstandard2.1 gets the portable algorithms that were already in
the tree as the fallback branches.

* BitOps is the single place the library reaches for BitOperations, so
  there is one file to audit instead of ninety-odd call sites. The wrappers
  are aggressively inlined, so the JIT still sees the intrinsic at the
  original call site.
* The 24 Bmi1/Bmi2/ArmBase sites keep their portable expression as the
  shared tail of the ternary, with only the "IsSupported ? intrinsic :"
  prefix conditional - so the two paths cannot drift apart.
* ReverseBits' table lookup uses Unsafe.Add to skip a bounds check that a
  byte index can never trip. Unsafe is a separate NuGet package on
  netstandard2.1, and this library takes no dependencies, so that target
  indexes normally instead.
* New RotateLeft/RotateRight/IsPow2 software fallbacks, tested against
  BitOperations like the existing ones.

Verifying the new target needed some thought: netstandard2.1 cannot host a
test project, and a normal run on .NET 8/10 only ever executes the
intrinsic branches, so the portable code would have shipped unexercised.
Building with -p:ForcePortable=true compiles the modern targets down the
portable branches instead, which runs the whole suite against the code
netstandard2.1 actually gets. CI now does both, and publishing is gated on
the portable run; it lives in its own job so a portable-compiled assembly
can never end up in the package.

All 430 tests pass on net8.0 and net10.0, both with intrinsics and forced
portable. Package now carries netstandard2.1, net6.0, net8.0 and net10.0
assets. Also fixes a CS0675 sign-extension warning in BitPacking.ReadBits.

+++ src/Tedd.BitUtils.Tests/BitPackingTest.cs
@@ -30,8 +30,8 @@ namespace Tedd.BitUtils.Tests
             for (var i = 0; i < bitCount; i++)
             {
                 var position = bitOffset + i;
-                var bit = (source[(int)(position >> 3)] >> (7 - (int)(position & 7))) & 1;
-                result = (result << 1) | (ulong)bit;
+                var bit = (uint)(source[(int)(position >> 3)] >> (7 - (int)(position & 7))) & 1u;
+                result = (result << 1) | bit;
             }
             return result;
         }
